requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
pytesseract
imgcat